Implement shared ride booking and matching module
Added full shared ride booking and matching functionality, including dual booking flows (AI_BOOKING and JOIN_RIDE), new controllers, services, DTOs, mappers, and repositories. Updated entities and enums to support new flows and statuses. Introduced configuration properties, business rules, and Flyway migrations for schema changes and performance indexes. Documentation for business flow, implementation plan, and implementation log included. Wallet integration, AI matching, pessimistic locking, and cancellation fee logic implemented for MVP.

+++ src/main/java/com/mssus/app/service/TransactionService.java
@@ -19,7 +19,7 @@ public interface TransactionService {
     void handleTopupFailed(String pspRef, String reason);
 
     // RIDE_HOLD flows
-    List<Transaction> createHold(Integer riderId, BigDecimal amount, Long bookingId, String description);
+    List<Transaction> createHold(Integer riderId, BigDecimal amount, Integer bookingId, String description);
 
     // RIDE_CAPTURE flows
     List<Transaction> captureFare(UUID groupId, Integer riderId, Integer driverId, BigDecimal totalFare,
